Split pads from 47BB8
