+++ config/splat.us.dra.yaml
@@ -205,6 +205,7 @@ segments:
       - [0x4493C, c, play]
       - [0x45D30, c, loading]
       - [0x47BB8, c, 47BB8]
+      - [0x48D24, c, pads]
       - [0x48EE4, c, save_mgr]
       - [0x4A538, c, 4A538]
       - [0x4F45C, c, collider]
