+++ src/dra/pads.c
@@ -1,0 +1,70 @@
+#include "dra.h"
+
+void ResetPadsRepeat(void) {
+    s8* ptr;
+    s32 i;
+
+    g_pads[0].repeat = 0;
+    ptr = g_PadsRepeatTimer;
+
+    for (i = 0; i < LEN(g_PadsRepeatTimer); i++) {
+        *ptr++ = 0x10;
+    }
+}
+
+void UpdatePadsRepeat(void) {
+    u16 button = 1;
+    u16 repeat = 0;
+    u16 unk = g_pads[0].tapped;
+    u16 pressed = g_pads[0].pressed;
+    u8* timers = g_PadsRepeatTimer;
+    s32 i = 0;
+
+    while (i < 0x10) {
+        if (pressed & button) {
+            if (unk & button) {
+                repeat |= button;
+                timers[0] = 0x10;
+            } else {
+                if (--timers[0] == 0xFF) {
+                    repeat |= button;
+                    timers[0] = 5;
+                }
+            }
+        }
+        i++;
+        timers++;
+        button <<= 1;
+    }
+    g_pads[0].repeat = repeat;
+}
+
+void InitializePads(void) {
+    Pad* pad;
+    s32 i;
+
+    PadInit(0);
+    for (pad = g_pads, i = 0; i < PAD_COUNT; i++, pad++) {
+        pad->tapped = 0;
+        pad->previous = 0;
+        pad->pressed = 0;
+    }
+    ResetPadsRepeat();
+}
+
+void ReadPads(void) {
+    Pad* pad;
+    s32 i;
+    u_long padd;
+
+    for (pad = g_pads, i = 0; i < PAD_COUNT; i++, pad++) {
+        pad->previous = pad->pressed;
+        padd = PadRead(i >> 1);
+        if (!(i & 1))
+            pad->pressed = padd;
+        else
+            pad->pressed = padd >> 0x10;
+        pad->tapped = (pad->pressed ^ pad->previous) & pad->pressed;
+    }
+    UpdatePadsRepeat();
+}
